+++ x/pricefeed/types/keys.go
@@ -1,6 +1,8 @@
 package types
 
-import sdk "github.com/cosmos/cosmos-sdk/types"
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
 
 const (
 	// ModuleName defines the module name
@@ -30,18 +32,13 @@ var (
 	// RawPriceFeedPrefix prefix for the raw pricefeed of an asset
 	RawPriceFeedPrefix = []byte{0x01}
 
-	// TWAPPrefix prefix for the current price of an asset
-	TWAPPrefix = []byte{0x02}
+	// Snapshot prefix for the median oracle price at a specific point in time
+	PriceSnapshotPrefix = []byte{0x03}
 )
 
 // CurrentPriceKey returns the prefix for the current price
 func CurrentPriceKey(pairID string) []byte {
 	return append(CurrentPricePrefix, []byte(pairID)...)
-}
-
-// CurrentTWAPKey returns the prefix for the current TWAP price
-func CurrentTWAPKey(twapPairID string) []byte {
-	return append(TWAPPrefix, []byte(twapPairID)...)
 }
 
 // RawPriceIteratorKey returns the prefix for the raw price for a single market
@@ -57,6 +54,16 @@ func RawPriceKey(pairID string, oracleAddr sdk.AccAddress) []byte {
 	return append(
 		RawPriceIteratorKey(pairID),
 		lengthPrefixWithByte(oracleAddr)...,
+	)
+}
+
+func PriceSnapshotKey(pairId string, blockHeight int64) []byte {
+	return append(
+		PriceSnapshotPrefix,
+		append(
+			[]byte(pairId),
+			sdk.Uint64ToBigEndian(uint64(blockHeight))...,
+		)...,
 	)
 }